First Commit - Added all neccessary files
A simple space invaders game that gives a new twist to and old classic.
The player has to use arrow keys to move a spaceship from left to right
and defeat five waves of alien enemies. The final level is a boss level
which the player has to beat in order to win the game